requirements.txt
requests
beautifulsoup4
openai
pytz
python-dotenv
